+++ esp-hal-common/src/rtc_cntl/rtc/esp32c2.rs
@@ -4,7 +4,6 @@ use crate::{
     clock::XtalClock,
     peripherals::{APB_CTRL, EXTMEM, RTC_CNTL, SPI0, SPI1, SYSTEM},
     regi2c_write_mask,
-    rom::rom_i2c_writeReg_Mask,
     rtc_cntl::{RtcCalSel, RtcClock, RtcFastClock, RtcSlowClock},
 };
 
@@ -29,10 +28,8 @@ const I2C_ULP_IR_FORCE_XPD_CK_LSB: u32 = 2;
 pub(crate) fn init() {
     let rtc_cntl = unsafe { &*RTC_CNTL::ptr() };
 
-    unsafe {
-        regi2c_write_mask!(I2C_DIG_REG, I2C_DIG_REG_XPD_DIG_REG, 0);
-        regi2c_write_mask!(I2C_DIG_REG, I2C_DIG_REG_XPD_RTC_REG, 0);
-    }
+    regi2c_write_mask!(I2C_DIG_REG, I2C_DIG_REG_XPD_DIG_REG, 0);
+    regi2c_write_mask!(I2C_DIG_REG, I2C_DIG_REG_XPD_RTC_REG, 0);
 
     unsafe {
         rtc_cntl
@@ -53,9 +50,9 @@ pub(crate) fn init() {
     unsafe {
         rtc_cntl.int_ena_rtc.write(|w| w.bits(0));
         rtc_cntl.int_clr_rtc.write(|w| w.bits(u32::MAX));
-
-        regi2c_write_mask!(I2C_ULP, I2C_ULP_IR_FORCE_XPD_CK, 0);
     }
+
+    regi2c_write_mask!(I2C_ULP, I2C_ULP_IR_FORCE_XPD_CK, 0);
 }
 
 pub(crate) fn configure_clock() {