+++ esp-hal-common/src/soc/esp32h2/mod.rs
@@ -12,4 +12,9 @@ pub(crate) mod constants {
 
     pub const I2S_DEFAULT_CLK_SRC: u8 = 1;
     pub const I2S_SCLK: u32 = 96_000_000;
+
+    pub const RMT_RAM_START: usize = 0x60007400;
+    pub const RMT_CHANNEL_RAM_SIZE: usize = 48;
+    pub const RMT_CLOCK_SRC: bool = false;
+    pub const RMT_CLOCK_SRC_FREQ: fugit::HertzU32 = fugit::HertzU32::MHz(32);
 }