+++ esp-hal-common/src/soc/esp32s2/mod.rs
@@ -9,4 +9,7 @@ pub mod ulp_core;
 pub(crate) mod constants {
     pub const I2S_SCLK: u32 = 160_000_000;
     pub const I2S_DEFAULT_CLK_SRC: u32 = 2;
+
+    pub const RMT_RAM_START: usize = 0x3f416400;
+    pub const RMT_CHANNEL_RAM_SIZE: usize = 64;
 }