+++ esp32c2-hal/examples/qspi_flash.rs
@@ -103,7 +103,7 @@ fn main() -> ! {
             zero_buf,
         )
         .unwrap();
-    (zero_buf, spi) = transfer.wait();
+    (zero_buf, spi) = transfer.wait().unwrap();
     delay.delay_ms(250u32);
 
     // erase sector
@@ -116,7 +116,7 @@ fn main() -> ! {
             zero_buf,
         )
         .unwrap();
-    (zero_buf, spi) = transfer.wait();
+    (zero_buf, spi) = transfer.wait().unwrap();
     delay.delay_ms(250u32);
 
     // write enable
@@ -129,7 +129,7 @@ fn main() -> ! {
             zero_buf,
         )
         .unwrap();
-    (_, spi) = transfer.wait();
+    (_, spi) = transfer.wait().unwrap();
     delay.delay_ms(250u32);
 
     // write data / program page
@@ -144,7 +144,7 @@ fn main() -> ! {
             send,
         )
         .unwrap();
-    (_, spi) = transfer.wait();
+    (_, spi) = transfer.wait().unwrap();
     delay.delay_ms(250u32);
 
     loop {
@@ -162,7 +162,7 @@ fn main() -> ! {
         // here we could do something else while DMA transfer is in progress
         // the buffers and spi is moved into the transfer and we can get it back via
         // `wait`
-        (receive, spi) = transfer.wait();
+        (receive, spi) = transfer.wait().unwrap();
 
         println!("{:x?}", &receive);
         for b in &mut receive.iter() {